changed error output and moved some functions
