+++ testAlgo.py
@@ -1,0 +1,23 @@
+import sys
+
+
+#!/usr/bin/env python3
+
+"""
+testAlgo.py
+
+Description: testing the wikipedia tree maker
+
+Usage:
+    python testAlgo.py [arguments]
+
+Author: Jake Brunning
+"""
+
+
+
+def main():
+    print("Hello, World!")
+
+if __name__ == "__main__":
+    main()
